fix: round exported pixel data values
- refactor pixel data export to use a pre-computed atom for exports.
- round exported values to 3 significant figures.
